refactored checks into separate files
