+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/tools/network-checks/apim-flows/checks/routeTableCheck.ts
@@ -1,0 +1,7 @@
+import { VirtualNetworkType } from '../contracts/APIMService';
+
+export function routeTableCheck(networkType: VirtualNetworkType) {
+    if (networkType != VirtualNetworkType.NONE) {
+    } else {
+    }
+}
